# docs/swagger.yaml
basePath: /api/v12
definitions:
...
  github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse:
    properties:
      error:
        example: Invalid JSON format
        type: string
    type: object
...
  github_com_QuentinRegnier_nubo-backend_internal_domain.RefreshMasterInput:
    properties:
      id_user:
        type: integer
      master_token:
        description: L'ancien MasterToken
        type: string
      username:
        description: Le username de l'utilisateur
        type: string
    required:
    - id_user
    - master_token
    - username
    type: object
  github_com_QuentinRegnier_nubo-backend_internal_domain.RefreshMasterResponse:
    properties:
      master_token:
        type: string
      message:
        type: string
      token:
        description: Le nouveau JWT
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Documentation de l'API.
    Pour mettre à jour :
  title: Mon API Propre
  version: "1.0"
paths:
  /auth/refresh-master:
    post:
      consumes:
      - application/json
      description: |
        Réinitialise toute la chaîne de sécurité (Ratchet, JWT, Secrets) en générant un nouveau MasterToken.
        Cette route est l'ultime recours ("Last Resort") lorsque le Ratchet est désynchronisé ou que le JWT est expiré depuis trop longtemps.

        **Mécanisme de Résilience :**
        1. Recherche la session via le MasterToken dans **Redis**.
        2. Si introuvable (crash cache), cherche dans **MongoDB**.
        3. Si introuvable, cherche dans **PostgreSQL** (Source de vérité).
        4. Si trouvé, valide la signature HMAC et réinitialise tout.

        **Actions Serveur :**
        * Génération de `NewMasterToken` et `NewJWT`.
        * Reset du Ratchet (Secret 0 = NewMaster, Secret 1 = DeviceToken).
        * Mise à jour asynchrone de Postgres et Mongo pour persister le nouveau MasterToken.
      parameters:
      - description: Bearer <Current_JWT> (Optionnel, pour continuité)
        in: header
        name: Authorization
        type: string
      - description: HMAC calculé avec l'ANCIEN MasterToken
        in: header
        name: X-Signature
        required: true
        type: string
      - description: Timestamp de la requête
        in: header
        name: X-Timestamp
        required: true
        type: string
      - description: Données de reset (MasterToken, UserID, Username)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.RefreshMasterInput'
      produces:
      - application/json
      responses:
        "200":
          description: Nouveaux identifiants générés
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.RefreshMasterResponse'
        "400":
          description: Format invalide ou Headers manquants
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
        "401":
          description: MasterToken introuvable ou Signature HMAC invalide
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
        "500":
          description: Erreur serveur critique (Génération/Sauvegarde)
          schema:
            $ref: '#/definitions/github_com_QuentinRegnier_nubo-backend_internal_domain.ErrorResponse'
      summary: Hard Refresh (Master Token Rotation)
      tags:
      - auth
...
swagger: "2.0"